refactor(digitsd): extract WebRTC callbacks into webrtc.go
Pure file-split refactor. No behavior change.

main.go shed 740 lines: the 2-party WebRTC entrypoints (InitiateCall,
AnswerCall, HangupCall), pipeline setup helpers (newPipeline,
startEncodeLoop), inbound RTP handler (remoteTrackHandler), ring-phase
pre-allocation (prepareAnswer, cleanupPreAnswer), and the ICE
connection-state machinery (triggerHangup, handleConnectionStateChange,
attemptICERestart, startRestartTimeout).

webrtc.go picks up the same code under package main with the same
receiver (*daemonCallbacks). Conference (mesh) callbacks
(MutePeer, MigrateToMesh, TearDownPeer, RequestConferenceMerge,
AddMeshPeer, RemoveMeshPeer, TearDownAllMeshPeers,
setupMeshResponder, currentPeer) stay on main.go for now; they get
their own file in the next refactor commit.

Imports trimmed: webrtc.go pulls only the packages it actually uses
(context, fmt, log/slog, time, audio, config, devmode, sigclient,
owebrtc, pion webrtc). main.go retains everything it still references.

Verification: go build, go test, golangci-lint, make build (Docker
arm64) all clean.
